+++ framework/db-local/src/test/java/is/codion/framework/db/local/DefaultLocalEntityConnectionTest.java
@@ -252,7 +252,8 @@ public class DefaultLocalEntityConnectionTest {
 						.build();
 
 		Entity.Key key = connection.insert(noPk);
-		assertTrue(key.isNull());
+		assertFalse(key.isNull());
+		assertFalse(key.primaryKey());
 	}
 
 	@Test